requirements.txt
Flask
flask_sqlalchemy
flask_login
PyMySQL
cryptography
mysql-connector-python
python-dotenv